+++ src/modules/private/shared/bankAccount/containers/Plaid.js
@@ -24,8 +24,8 @@ export default class Plaid extends Component {
     }
     return (
       <div>
-        <Header as="h1" textAlign="center">Link Bank Account</Header>
-        <Header as="h4" textAlign="center">Select your bank from the list</Header>
+        <Header as="h3" textAlign="center">Link Bank Account</Header>
+        <Header as="h6" textAlign="center">Select your bank from the list</Header>
         <Form>
           <Input
             fluid
@@ -87,7 +87,7 @@ export default class Plaid extends Component {
           }
         </div>
         <div className="center-align">
-          <Button className="theme-link" primary onClick={() => this.props.bankAccountStore.setBankLinkInterface('form')}>or enter it manually</Button>
+          <Button className="cancel-link" onClick={() => this.props.bankAccountStore.setBankLinkInterface('form')}>or enter it manually</Button>
         </div>
       </div>
     );