+++ src/theme/form/src/FormCheckbox.js
@@ -21,7 +21,7 @@ const FormCheckbox = observer((props) => {
                 label={
                   <label>{c.label}
                     {c.tooltip &&
-                      <Popup trigger={<Icon name="help circle outline" />} content={c.tooltip} position="top center" />
+                      <Popup trigger={<Icon name="help circle" />} content={c.tooltip} position="top center" />
                     }
                   </label>
                 }
@@ -37,7 +37,7 @@ const FormCheckbox = observer((props) => {
                   {c.label}
                   {tooltip &&
                     <Popup
-                      trigger={<Icon name="help circle outline" />}
+                      trigger={<Icon name="help circle" />}
                       content={tooltip}
                       position="top center"
                       className="center-align"